requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai-whisper==20230314
language-tool-python==2.7.1
setuptools>=65.0.0
wheel
torch>=2.0.0
torchaudio>=2.0.0
